Verify if authAccount trying update a recipeStep is the Author

+++ src/main/java/com/leorizick/recipeapp/services/domain/service/recipe/RecipeStepCrud.java
@@ -2,6 +2,8 @@ package com.leorizick.recipeapp.services.domain.service.recipe;
 
 import com.leorizick.recipeapp.entities.recipe.RecipeStep;
 import com.leorizick.recipeapp.repositories.RecipeStepsRepository;
+import com.leorizick.recipeapp.services.domain.service.config.auth.AuthenticationContext;
+import com.leorizick.recipeapp.services.exceptions.AccountTypeNotAllowed;
 import com.leorizick.recipeapp.services.exceptions.NotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -13,6 +15,7 @@ import javax.transaction.Transactional;
 public class RecipeStepCrud {
 
     private final RecipeStepsRepository recipeStepsRepository;
+    private final AuthenticationContext authenticationContext;
 
     public RecipeStep findById(Long id) {
         return recipeStepsRepository.findById(id)
@@ -20,15 +23,23 @@ public class RecipeStepCrud {
     }
     @Transactional
     public RecipeStep save(RecipeStep recipeStep){
+        verifyAuthorIdAndAuthAccountId(recipeStep);
         return recipeStepsRepository.save(recipeStep);
     }
 
     public void delete(Long id){
         RecipeStep recipeStep = findById(id);
+        verifyAuthorIdAndAuthAccountId(recipeStep);
         recipeStepsRepository.deleteById(id);
     }
 
     public void DeleteByRecipeId(Long recipeId){
         recipeStepsRepository.deleteByRecipeId(recipeId);
+    }
+
+    private void verifyAuthorIdAndAuthAccountId(RecipeStep recipeStep){
+        if(!authenticationContext.getAccountId().equals(recipeStep.getRecipe().getAuthor().getId())){
+            throw new AccountTypeNotAllowed("author");
+        }
     }
 }
